refactor: remove model folder
