+++ service/hub/internal/server/handler/handler.go
@@ -1,6 +1,8 @@
 package handler
 
 import (
+	"time"
+
 	rabbitmq "github.com/rabbitmq/amqp091-go"
 	"gorm.io/gorm"
 )
@@ -19,4 +21,23 @@ type Response struct {
 	Total  int64  `json:"total"`
 	Cursor string `json:"cursor,omitempty"`
 	Result any    `json:"result"`
+}
+
+type GetRequest struct {
+	Address   string    `param:"address"`
+	Limit     int       `query:"limit"`
+	Cursor    string    `query:"cursor"`
+	Type      []string  `query:"type"`
+	Tag       string    `query:"tag"`
+	Network   []string  `query:"network"`
+	Platform  []string  `query:"platform"`
+	Timestamp time.Time `query:"timestamp"`
+	Hash      string    `query:"hash"`
+}
+
+type GetExchangeRequest struct {
+	ExchangeType string   `param:"exchange_type"`
+	Cursor       int      `query:"cursor"`
+	Name         []string `query:"name"`
+	Network      []string `query:"network"`
 }
